Working with File Upoad, Add Invoice

+++ controllers/admin.js
@@ -1,6 +1,9 @@
 const Product = require('../models/product');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check')
+
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
     return res.redirect('/login')
@@ -17,9 +20,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+    });
+  }
+
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     console.log(errors.array())
@@ -38,6 +57,9 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price, 
@@ -89,7 +111,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -101,7 +123,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true, 
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -115,10 +136,14 @@ exports.postEditProduct = (req, res, next) => {
     if (product.userId.toString() !== req.user._id.toString()) {
       return res.redirect('/');
     }
+  
     product.title = updatedTitle;
     product.price = updatedPrice;
     product.description = updatedDesc;
-    product.imageUrl = updatedImageUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl)
+      product.imageUrl = image.path;
+    }
     return product.save()
     .then(result => {
       console.log('UPDATED PRODUCT!');
@@ -151,14 +176,20 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('product not found'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({_id: prodId, userId: req.user._id})
+  })
+  .then(() => {
       console.log('Delete PRODUCT');
       res.redirect('/admin/products');
     })
-    .catch(err => {
+  .catch(err => {
       const error = new Error('err');
       error.httpStatusCode = 500;
       return next(error);
-    });
+  });
 };
